Implement triplicate exercise
From https://clojure.org/guides/learn/functions

Note that multiple expressions in function body should not be wrapped in
parens, unles they should be further evaluated. If wrapped, the value of
the first expression (nil in this case) will be called as function with
values of subsequent expressions as arguments, as in:

    (nil nil nil)

This leads to a dreaded null pointer exception.