Refactor math exercise to use map
Vector of math expressions is mapped to a function that concatenates
them with their value. The result is printed.

Note that map is lazy, so printing must be outside of the iterator
function. Otherwis nothing is going to happen.

I'm not sure if this is more readable. It's defienitely more concise.